requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
cssselect==1.1.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
h11==0.13.0
helium==3.0.8
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.0
multidict==6.0.2
numpy==1.23.0
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.3
parse==1.19.0
polygon-api-client==1.1.0
py-cord==2.0.0rc1
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-html==0.10.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
webdriver-manager==3.7.1
webencodings==0.5.1
websockets==10.3
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
